Fix EditButton logic
List editing should be allow if either the user is a Member of the
Inner Circle, or if the user is a Sender for the list. Previously,
Inner Circle membership was only checked if the user had no Role for
the list. Specifically, if the user was a Receiver for the list, Inner
Circle membership would be ignored, and list editing would be
forbidden.

Signed-off-by: Maximilian Marx <mmarx@wh2.tu-dresden.de>